requirements.txt
alembic==1.0.8
asn1crypto==0.24.0
Babel==2.6.0
blinker==1.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
config==0.4.0
cryptography==2.6.1
dominate==2.3.5
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-PyMongo==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
httpie==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jwt==0.6.1
Mako==1.0.7
MarkupSafe==1.1.1
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pymodm==0.4.1
pymongo==3.7.2
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
redis==3.2.1
requests==2.21.0
rq==0.13.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
